song-vocab/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
ollama==0.1.6
instructor==0.4.5
pydantic==2.6.1
python-dotenv==1.0.1
google-search-results==2.4.2
beautifulsoup4==4.12.3
requests==2.31.0
